moved blender-specific (exr) postprocessing to blender_utils.py
